refactor duplicated functions
